Technical/Refactor dnsResolver#mxRecords

+++ dns_resolver_test.go
@@ -142,19 +142,19 @@ func TestDnsResolverMxRecords(t *testing.T) {
 	domain := randomDomain()
 
 	t.Run("when target MX record found", func(t *testing.T) {
-		mxPriorityFirst, mxFirst, mxPrioritySecond, mxSecond := 20, randomDomain(), 10, randomDomain()
+		mxPriorityFirst, mxFirst, mxPrioritySecond, mxSecond := uint16(20), randomDomain(), uint16(10), randomDomain()
 		dnsRecords := map[string]mockdns.Zone{
 			domain + ".": {
 				MX: []net.MX{
-					{Host: mxFirst, Pref: uint16(mxPriorityFirst)},
-					{Host: mxSecond, Pref: uint16(mxPrioritySecond)},
+					{Host: mxFirst, Pref: mxPriorityFirst},
+					{Host: mxSecond, Pref: mxPrioritySecond},
 				},
 			},
 		}
 		dnsResolver := createDnsResolver(dnsRecords)
 		resolvedMxPriorities, resolvedMxHostNames, err := dnsResolver.mxRecords(domain)
 
-		assert.Equal(t, []int{mxPrioritySecond, mxPriorityFirst}, resolvedMxPriorities)
+		assert.Equal(t, []uint16{mxPrioritySecond, mxPriorityFirst}, resolvedMxPriorities)
 		assert.Equal(t, []string{mxSecond, mxFirst}, resolvedMxHostNames)
 		assert.Nil(t, err)
 	})
